Inline transact validation into the create and update processors
Each processor now specializes its private-tx-hash inputs directly:
create nullifies the address UTXO and binds its hash, update spends the
previous state UTXO with the zero address slot. Drops validate_transact
and the Transition struct from instructions/shared.rs; private_tx_hash
stays shared.
