+++ sdk-tests/compression/program/src/instructions/shared.rs
@@ -9,33 +9,13 @@ use pinocchio::{
 use solana_address::address;
 use zolana_account_checks::AccountIterator;
 use zolana_hasher::{hash_chain::create_hash_chain_from_slice, Hasher, Poseidon};
-use zolana_interface::{
-    event::MessageData,
-    instruction::{
-        instruction_data::transact::{
-            CircuitId, ExternalDataHash, OwnerTag, ResolvedOutput, TransactIxData,
-        },
-        tag::TRANSACT,
-    },
-    N_PUBLIC_SLOTS,
-};
+#[cfg(any(target_os = "solana", target_arch = "bpf"))]
+use zolana_interface::instruction::tag::TRANSACT;
 
-use crate::{
-    error::CompressionError,
-    state::{
-        derive_address, derive_blinding, derive_state, nullifier, plaintext_payload,
-        state_utxo_hash,
-    },
-};
+use crate::error::CompressionError;
 
 pub const DEFAULT_TREE: Address = address!("trEEbaNobcTESNmtsPBj3FX27q5sDCQePV2kb12FYho");
 pub const SPP_PROGRAM: Address = address!("sppXZU59VoYodv9Accs4hHNTjYiuYmDFyFVjUjPxFsG");
-
-pub struct Transition {
-    pub old: Option<(u64, [u8; 32])>,
-    pub new_value: u64,
-    pub output_seed: [u8; 32],
-}
 
 #[cfg(any(target_os = "solana", target_arch = "bpf"))]
 pub fn derive_pda(authority: &Address) -> (Address, u8) {
@@ -105,98 +85,7 @@ impl<'a> TransitionAccounts<'a> {
     }
 }
 
-pub fn validate_transact(
-    authority: &Address,
-    pda: &Address,
-    transition: &Transition,
-    transact: &TransactIxData,
-) -> ProgramResult {
-    if transact.expiry_unix_ts != u64::MAX
-        || transact.circuit != CircuitId::ConfidentialEddsa(1, 1, N_PUBLIC_SLOTS as u8)
-        || transact.tx_viewing_pk != [0u8; 33]
-        || transact.salt != [0u8; 16]
-        || !transact.interface_transfers.is_empty()
-        || transact.data_hash.is_some()
-        || transact.ring_data_hash.is_some()
-        || !transact.messages.is_empty()
-    {
-        return Err(CompressionError::InvalidTransact.into());
-    }
-    let [input] = transact.inputs.as_slice() else {
-        return Err(CompressionError::InvalidTransact.into());
-    };
-    let [output] = transact.outputs.as_slice() else {
-        return Err(CompressionError::InvalidTransact.into());
-    };
-    let Some(output_data) = output.data.as_deref() else {
-        return Err(CompressionError::InvalidTransact.into());
-    };
-
-    let pda_bytes = pda.to_bytes();
-    let address = derive_address(&pda_bytes)?;
-    let state = derive_state(&address.address, authority.as_array(), transition.new_value)?;
-    let output_blinding = derive_blinding(&transition.output_seed)?;
-    let expected_output_hash =
-        state_utxo_hash(&address.owner_hash, &state.data_hash, &output_blinding)?;
-    if output.utxo_hash != expected_output_hash || output.owner_tag != OwnerTag::Inline(pda_bytes) {
-        return Err(CompressionError::InvalidState.into());
-    }
-    let expected_payload =
-        plaintext_payload(&pda_bytes, &state.state_data, transition.output_seed)?;
-    if output_data != expected_payload.as_slice() {
-        return Err(CompressionError::InvalidState.into());
-    }
-
-    let resolved_output = [ResolvedOutput {
-        utxo_hash: &output.utxo_hash,
-        owner_tag: pda_bytes,
-        data: Some(output_data),
-    }];
-    let messages: &[MessageData] = &[];
-    let external_data_hash = ExternalDataHash {
-        spp_instruction_discriminator: TRANSACT,
-        expiry_unix_ts: transact.expiry_unix_ts,
-        interface_transfers: &[],
-        data_hash: None,
-        ring_data_hash: None,
-        tx_viewing_pk: &transact.tx_viewing_pk,
-        salt: &transact.salt,
-        outputs: &resolved_output,
-        messages,
-    }
-    .hash()
-    .map_err(|_| CompressionError::HashingFailed)?;
-
-    let (input_hash, address_hash, expected_nullifier) = match transition.old {
-        None => (
-            [0u8; 32],
-            address.address_utxo_hash,
-            nullifier(&address.address_utxo_hash, &address.address_seed)?,
-        ),
-        Some((old_value, old_blinding)) => {
-            let old_state = derive_state(&address.address, authority.as_array(), old_value)?;
-            let old_hash =
-                state_utxo_hash(&address.owner_hash, &old_state.data_hash, &old_blinding)?;
-            (old_hash, [0u8; 32], nullifier(&old_hash, &old_blinding)?)
-        }
-    };
-
-    if input.nullifier_hash != expected_nullifier {
-        return Err(CompressionError::InvalidAddress.into());
-    }
-    let expected_private_tx = private_tx_hash(
-        input_hash,
-        expected_output_hash,
-        address_hash,
-        &external_data_hash,
-    )?;
-    if transact.private_tx_hash != expected_private_tx {
-        return Err(CompressionError::InvalidTransact.into());
-    }
-    Ok(())
-}
-
-fn private_tx_hash(
+pub fn private_tx_hash(
     input_hash: [u8; 32],
     output_hash: [u8; 32],
     address_hash: [u8; 32],
